style: Detach the modal folder in the subfolder of movies
modal은 재사용이 가능하므로 movies 하위 폴더인 modal 폴더를 component의 하위폴더로 이동하였다
